requirements.txt
flask>=3.1.3
flask-cors>=6.0.2
flask-sqlalchemy>=3.1.1
bcrypt>=5.0.0
pyjwt>=2.12.1
